+++ include/Sequencer/RTPScene.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include "RTPEventNoteSequence.h"
 #include "Structs.h"
 #include <vector>
@@ -9,45 +10,45 @@ using namespace std;
 class RTPScene{
   vector<unique_ptr<RTPEventNoteSequence>> SequencerScene;
   String _name;
-  int _NSequences;
-  int _selectedSequence;
+  uint8_t _NSequences;
+  uint8_t _selectedSequence;
   NotesPlayer& _notesPlayer; 
   MusicManager& _musicManager;
   RTPSequenceNoteStates _seqStates;
 public:
-  RTPScene(String name, int NSequences, int scene, NotesPlayer& notesPlayer, MusicManager& musicManager);
+  RTPScene(String name, uint8_t NSequences, uint8_t scene, NotesPlayer& notesPlayer, MusicManager& musicManager);
   void playScene();
   void fordwardScene();
   void backwardScene();
   void resetScene();
-  void setSelectedSequence(int selectedSequence);
-  int getSelectedSequence();
-  int getSelectedSequencePageOffset();
-  int getSelectedSequencePage();
-  int getSelectedSequenceMidiChannel();
-  int getSize();
+  void setSelectedSequence(uint8_t selectedSequence);
+  uint8_t getSelectedSequence();
+  uint16_t getSelectedSequencePageOffset();
+  uint8_t getSelectedSequencePage();
+  uint8_t getSelectedSequenceMidiChannel();
+  uint8_t getSize();
   bool isSelectedSequenceRecording();
   void toggleSelectedSequenceRecording();
   RTPSequenceNoteStates getSelectedSequenceNoteStates();
   RTPSequencesState getSequencesState();
   SequenceSettings getSelectedSequenceSettings();
-  void selectParameterInSequence(int parameterIndex);
+  void selectParameterInSequence(uint8_t parameterIndex);
   void incselectParameterInSequence();
   void decselectParameterInSequence();
   int getSelectedParameterInSequenceValue();
   String getSelectedParameterInSequenceName();
   void incselectPageInSequence();
   void decselectPageInSequence();
-  int getSelectedSequenceCurrentPosition();
-  int getSequenceType(int sequenceIndex);
-  void toggleSequence(int sequenceIndex);
-  void toggleNoteInSequence(int position);
+  uint16_t getSelectedSequenceCurrentPosition();
+  uint8_t getSequenceType(uint8_t sequenceIndex);
+  void toggleSequence(uint8_t sequenceIndex);
+  void toggleNoteInSequence(uint16_t position);
   void editNoteInCurrentPosition(ControlCommand command);
   RTPSequenceNoteStates getSequenceNoteStates();
-  int getSequenceColor();
+  uint32_t getSequenceColor();
   void dumpSequencesToJson();
   
   // For persistence manager
-  RTPEventNoteSequence* getSequence(int index);
+  RTPEventNoteSequence* getSequence(uint8_t index);
   const RTPEventNoteSequence* getSequence(int index) const;
 };
