+++ include/Sequencer/RTPSequencer.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <RTPScene.h>
 #include "NotesPlayer.hpp"
 #include "Structs.h"
@@ -9,44 +10,44 @@ using namespace std;
 
 class RTPSequencer{
   vector<RTPScene*> Sequencer; // Change to pointer type
-  int _NScenes;
-  int _selectedScene;
+  uint8_t _NScenes;
+  uint8_t _selectedScene;
   NotesPlayer _notesPlayer;
   MusicManager& _musicManager;
 
 public:
-  RTPSequencer(int NScenes, MusicManager& musicManager);
+  RTPSequencer(uint8_t NScenes, MusicManager& musicManager);
   ~RTPSequencer(); // Destructor for cleaning up
   void playAndMoveSequencer();
   void stopAndCleanSequencer();
   void pauseSequencer();
-  int getSelectedSequencePosition();
-  int getSelectedSequencePageOffset();
-  int getSelectedSequencePage();
+  uint16_t getSelectedSequencePosition();
+  uint16_t getSelectedSequencePageOffset();
+  uint8_t getSelectedSequencePage();
   bool isSelectedSequenceRecording();
   void toggleSelectedSequenceRecording();
   SequenceSettings getSelectedSequenceSettings();
-  void selectScene(int scene);
+  void selectScene(uint8_t scene);
   void increaseSelectedScene();
   void decreaseSelectedScene();
-  int getSelectScene();
-  int getSelectedSequence();
-  int getSelectedSequenceMidiChannel();
+  uint8_t getSelectScene();
+  uint8_t getSelectedSequence();
+  uint8_t getSelectedSequenceMidiChannel();
   void addScene(RTPScene* scene); // Change to pointer type
-  void removeScene(int scene);
-  void selectParameterInSequence(int parameterIndex);
+  void removeScene(uint8_t scene);
+  void selectParameterInSequence(uint8_t parameterIndex);
   void incSelectParameterInSequence();
   void decSelectParameterInSequence();
   int getSelectedParameterInSequenceValue();
   String getSelectedParameterInSequenceName();
-  void toggleSequence(int sequenceIndex);
+  void toggleSequence(uint8_t sequenceIndex);
   RTPSequencesState getSequencesState();
-  void selectSequence(int sequenceIndex);
+  void selectSequence(uint8_t sequenceIndex);
   RTPSequenceNoteStates getSelectedSequenceNoteStates();
-  void toggleNoteInSceneInSelectedSequence(int position);
+  void toggleNoteInSceneInSelectedSequence(uint16_t position);
   void nudgePageInSelectedSequence(ControlCommand command);
   void editNoteInCurrentPosition(ControlCommand command);
-  int getSelectedSequenceColor();
+  uint32_t getSelectedSequenceColor();
   void dumpSequencesToJson();
   
   // For persistence manager
