+++ src/Sequencer/RTPEventNote.cpp
@@ -1,13 +1,13 @@
 #include "RTPEventNote.hpp"
 
-RTPEventNote::RTPEventNote(bool state, int note){
+RTPEventNote::RTPEventNote(bool state, uint8_t note){
 	_state = state;
 	_note = note;
 	_velocity = 100;
 	_read = 0;
 }
 
-RTPEventNote::RTPEventNote(bool state, int note, int velocity){
+RTPEventNote::RTPEventNote(bool state, uint8_t note, uint8_t velocity){
 	_state = state;
 	_note = note;
 	_velocity = velocity;
@@ -23,30 +23,30 @@ bool RTPEventNote::eventState() const {
 	return _state;
 }
 
-int RTPEventNote::getEventNote(){
+uint8_t RTPEventNote::getEventNote(){
 	return _note;
 }
 
 // Const version for JSON serialization
-int RTPEventNote::getEventNote() const {
+uint8_t RTPEventNote::getEventNote() const {
 	return _note;
 }
 
-int RTPEventNote::getEventRead(){
+uint8_t RTPEventNote::getEventRead(){
 	return _read;
 }
 
 // Const version for JSON serialization
-int RTPEventNote::getEventRead() const {
+uint8_t RTPEventNote::getEventRead() const {
 	return _read;
 }
 
-int RTPEventNote::getEventVelocity(){
+uint8_t RTPEventNote::getEventVelocity(){
 	return _velocity;
 }
 
 // Const version for JSON serialization
-int RTPEventNote::getEventVelocity() const {
+uint8_t RTPEventNote::getEventVelocity() const {
 	return _velocity;
 }
 
@@ -54,15 +54,15 @@ void RTPEventNote::setEventState(bool state){
 	_state = state;
 }
 
-void RTPEventNote::setEventNote(int note){
+void RTPEventNote::setEventNote(uint8_t note){
 	_note = note;
 }
 
-void RTPEventNote::setEventRead(int read){
+void RTPEventNote::setEventRead(uint8_t read){
 	_read = read;
 }
 
-void RTPEventNote::setEventVelocity(int velocity){
+void RTPEventNote::setEventVelocity(uint8_t velocity){
 	_velocity = velocity;
 }
